SAM-1050
git-svn-id: https://source.sakaiproject.org/svn/sam/trunk@84422 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ samigo/samigo-app/src/java/org/sakaiproject/tool/assessment/ui/listener/author/ItemAddListener.java
@@ -1361,7 +1361,7 @@ public class ItemAddListener
 					log.debug("oneLabel = " + oneLabel);
 					answer.setText(oneAnswer);
 					answer.setLabel(oneLabel);
-					if (choiceSequence == matchSequence) {
+					if (choiceSequence.longValue() == matchSequence.longValue()) {
 						answer.setIsCorrect(Boolean.TRUE);
 					} else {
 						answer.setIsCorrect(Boolean.FALSE);
